refactor getDigitalOceanRegions and flattenRegion into separate file
