+++ digitalocean/datasource_digitalocean_regions.go
@@ -1,7 +1,6 @@
 package digitalocean
 
 import (
-	"context"
 	"fmt"
 	"sort"
 	"strconv"
@@ -106,61 +105,6 @@ func dataSourceDigitalOceanRegionsRead(d *schema.ResourceData, meta interface{})
 	}
 
 	return nil
-}
-
-func getDigitalOceanRegions(client *godo.Client) (map[string]godo.Region, error) {
-	allRegions := map[string]godo.Region{}
-
-	opts := &godo.ListOptions{
-		Page:    1,
-		PerPage: 200,
-	}
-
-	for {
-		regions, resp, err := client.Regions.List(context.Background(), opts)
-
-		if err != nil {
-			return nil, fmt.Errorf("Error retrieving regions: %s", err)
-		}
-
-		for _, region := range regions {
-			allRegions[region.Slug] = region
-		}
-
-		if resp.Links == nil || resp.Links.IsLastPage() {
-			break
-		}
-
-		page, err := resp.Links.CurrentPage()
-		if err != nil {
-			return nil, fmt.Errorf("Error retrieving regions: %s", err)
-		}
-
-		opts.Page = page + 1
-	}
-
-	return allRegions, nil
-}
-
-func flattenRegion(region godo.Region) map[string]interface{} {
-	flattenedRegion := map[string]interface{}{}
-	flattenedRegion["slug"] = region.Slug
-	flattenedRegion["name"] = region.Name
-	flattenedRegion["available"] = region.Available
-
-	sizesSet := schema.NewSet(schema.HashString, []interface{}{})
-	for _, size := range region.Sizes {
-		sizesSet.Add(size)
-	}
-	flattenedRegion["sizes"] = sizesSet
-
-	featuresSet := schema.NewSet(schema.HashString, []interface{}{})
-	for _, feature := range region.Features {
-		featuresSet.Add(feature)
-	}
-	flattenedRegion["features"] = featuresSet
-
-	return flattenedRegion
 }
 
 func filterDigitalOceanRegions(regions []godo.Region, filters []commonFilter) []godo.Region {
